REFACTOR :  move POST add-subscription to subscription module
